+++ src/main/java/com/tsp/new_tsp_admin/api/domain/production/AdminProductionDTO.java
@@ -19,8 +19,8 @@ import java.util.List;
 @AllArgsConstructor
 @ApiModel(value = "프로덕션 관련 변수")
 public class AdminProductionDTO extends NewCommonDTO {
-    @ApiModelProperty(required = true, value = "rnum", hidden = true, example = "1")
-    private Integer rnum;
+    @ApiModelProperty(required = true, value = "rowNum", hidden = true, example = "1")
+    private Integer rowNum;
 
     @ApiModelProperty(required = true, value = "idx", hidden = true, example = "1")
     private Long idx;
